replaced write()
