+++ plugin/taskscanner.lua
@@ -2,9 +2,9 @@ vim.api.nvim_create_user_command("WriteTasks", function()
   require("taskscanner").write_tasks()
 end, {})
 
-vim.api.nvim_create_user_command("WriteNewTasks", function()
-  require("taskscanner").write_new_tasks()
-end, {})
+-- vim.api.nvim_create_user_command("WriteNewTasks", function()
+--   require("taskscanner").write_tasks()
+-- end, {})
 
 
 vim.api.nvim_create_user_command("SyncTasks", function()
