+++ tests/fixtures/current_tasks.md
@@ -1,0 +1,26 @@
+## Urgent
+- [ ] Important Task 1
+- [ ] Prepare for meeting
+- [ ] Reply to lawyer
+
+### Pdi
+- [ ] prepare launch
+
+## Ai
+- [ ] Some task related to AI
+- [ ] exciting new task
+
+## Dev
+- [ ] aaa item
+- [ ] zzz item
+
+## Work
+- [ ] fix bug in module
+- [ ] show this instead
+
+### BackendTeam
+- [ ] refactor database
+
+## Untagged
+- [ ] Normal task 2
+- [ ] buy groceries
