+++ tests/write_spec.lua
@@ -1,112 +1,104 @@
 ---@diagnostic disable: undefined-global, undefined-field
 require("tests.test_helper")
 
+package.loaded["configs.notes"] = {
+  notes_dir = "tests/fixtures",
+}
+
+local write = require("taskscanner.write")
+local util = require("taskscanner.util")
+
+local notes_dir = "tests/fixtures"
+local output_file = notes_dir .. "/current_tasks.md"
+
+local function write_file(path, lines)
+  assert(type(lines) == "table", "lines must be a table of strings")
+  local f = assert(io.open(path, "w"))
+  f:write(table.concat(lines, "\n"))
+  f:close()
+end
+
+local function read_file(path)
+  local f = assert(io.open(path, "r"))
+  local content = f:read("*a")
+  f:close()
+  return vim.split(content, "\n")
+end
+
 describe("write_tasks", function()
-  local notes_dir = "tests/fixtures"
-  local output_file = notes_dir .. "/current_tasks.md"
-  local write
-
   before_each(function()
-    local modules = reset_taskscanner_modules(notes_dir)
-    write = modules.write
+    -- Clean slate
+    write_file(output_file, {})
 
-    local f = io.open(output_file, "w")
-    if f then
-      f:close()
-    end
+    -- Sample task input
+    write_file(notes_dir .. "/tasks1.md", {
+      "- [ ] #task fix bug in module #work",
+      "- [ ] #task refactor database #work #backend_team",
+      "- [ ] #task buy groceries",            -- untagged
+      "- [ ] #task prepare launch #urgent #pdi",
+      "- [X] #task old completed task #work", -- should be ignored
+      "- [ ] #task exciting new task #ai",    -- should be ignored
+
+    })
   end)
 
-  after_each(function()
-    os.remove(output_file)
+  it("writes structured tasks grouped by tags", function()
+    write.write_tasks('tests/fixtures/tasks1.md')
+
+    local lines = read_file(output_file)
+
+    assert.same({
+      "## Urgent",
+      "",
+      "### Pdi",
+      "- [ ] prepare launch",
+      "",
+      "## Ai",
+      "- [ ] exciting new task",
+      "",
+      "## Work",
+      "- [ ] fix bug in module",
+      "",
+      "### BackendTeam",
+      "- [ ] refactor database",
+      "",
+      "## Untagged",
+      "- [ ] buy groceries",
+      "",
+    }, lines)
   end)
 
-  it("returns source file mappings for each task", function()
-    local ok, result = pcall(function()
-      return write.write_tasks()
-    end)
+  it("omits completed tasks", function()
+    write_file(notes_dir .. "/tasks2.md", {
+      "- [X] #task do not show this #work",
+      "- [ ] #task show this instead #work",
+    })
+    write.write_tasks()
+    local lines = read_file(output_file)
 
-    -- print("RAW PCALL RETURN:", ok, result)
-    assert.are.same(true, ok) -- more strict than is_true
-    assert.is_table(result)
-
-    for task, path in pairs(result) do
-      -- print(string.format("[✓] %s -> %s", task, path))
-      assert.are.same(true, ok)
-      assert.is_table(result)
-    end
-  end)
-
-  it("writes urgent and normal tasks to current_tasks.md", function()
-    local sources = write.write_tasks()
-
-    local f = io.open(output_file, "r")
-    assert.is_not_nil(f)
-
-    local content = f:read("*all")
-    f:close()
-
-    -- Confirm both headers are present
-    assert.is_true(content:find("## Urgent Tasks") ~= nil)
-    assert.is_true(content:find("## Tasks") ~= nil)
-
-    -- Confirm #urgent tasks appear under correct section
-    assert.is_true(content:find("#urgent") ~= nil)
-    assert.is_true(content:find("#ai") ~= nil)
-  end)
-
-  it("deduplicates tasks and skips completed ones", function()
-    local sources = write.write_tasks()
-
-    local f = io.open(output_file, "r")
-    local lines = {}
-    for line in f:lines() do
-      table.insert(lines, line)
-    end
-    f:close()
-
-    -- Check that all lines are unique (no deduplication issues)
-    local seen = {}
     for _, line in ipairs(lines) do
-      if line:match("^%- %[ %]") then
-        assert.is_nil(seen[line], "Duplicate task found: " .. line)
-        seen[line] = true
+      assert.not_equal("- [ ] do not show this", line)
+    end
+  end)
+
+  it("sorts alphabetically within headers", function()
+    write_file(notes_dir .. "/tasks3.md", {
+      "- [ ] #task zzz item #dev",
+      "- [ ] #task aaa item #dev",
+    })
+    write.write_tasks()
+    local lines = read_file(output_file)
+
+    local dev_start
+    for i, line in ipairs(lines) do
+      if line == "## Dev" then
+        dev_start = i
+        break
       end
     end
 
-    -- Check that completed task doesn't appear
-    for _, line in ipairs(lines) do
-      assert.is_nil(line:match("^%- %[X%]"), "Completed task should not be written: " .. line)
-    end
-  end)
-
-  it("returns source file mappings for each task", function()
-    local ok, result = pcall(function()
-      return write.write_tasks()
-    end)
-
-    if not ok then
-      print("❌ write_tasks threw an error:")
-      print(result)
-    end
-
-    assert.are.same(true, ok)
-    assert.is_table(result)
-
-    if not next(result) then
-      print("⚠️ No tasks found in sources")
-    end
-
-    for task, path in pairs(result) do
-      assert.is_string(task)
-      assert.is_string(path)
-      assert.is_truthy(path:match("%.md$"))
-    end
-  end)
-
-  it("handles missing or unreadable notes directory gracefully", function()
-    package.loaded["configs.notes"] = { notes_dir = "nonexistent/path" }
-    local ok, result = pcall(write.write_tasks)
-    assert.is_true(ok)
-    assert.are.same({}, result)
+    assert.is_truthy(dev_start)
+    assert.equal("- [ ] aaa item", lines[dev_start + 1])
+    assert.equal("- [ ] zzz item", lines[dev_start + 2])
   end)
 end)
